Switch dom-delegate for ftdomdelegate.
All other maintained components use ftdomdelegate, which is the
same except for the name. We don't want two versions included.

We can do this now @i-like-robots has removed the  "dom-delegate"
name from ftdomdelegate and published to npm:
https://github.com/Financial-Times/ftdomdelegate/pull/93

This would usually be a breaking change, but I think given the release date of
this component it's acceptable to treat as a patch. Thoughts?